+++ src/backend/config.ts
@@ -1,21 +1,27 @@
 import { config } from 'dotenv';
 import { cleanEnv, str } from 'envalid';
+import fetch from 'node-fetch';
 import { UserData } from '../types/backend';
 import { Region, Server } from '../types/riot';
 import { slugify } from './utils';
 
-config();
-
 interface Config {
     REGION: string;
     DATADRAGON_VERSION: string;
+    LOCALE: string;
     RIOT_API_KEY: string;
     USERS: string;
 }
 
+config();
+
+let latestDataDragonVersion: string = '';
+
+
 export const env = cleanEnv<Config>(process.env, {
     REGION: str({ choices: ['americas', 'asia', 'europe'] }),
     DATADRAGON_VERSION: str(),
+    LOCALE: str(),
     RIOT_API_KEY: str(),
     USERS: str(),
 });
@@ -53,5 +59,18 @@ export function riotAPIKey(): string {
 }
 
 export function datadragonVersion(): string {
-    return env.DATADRAGON_VERSION;
+    return env.DATADRAGON_VERSION === 'latest' ? latestDataDragonVersion : env.DATADRAGON_VERSION;
+}
+export function locale(): string {
+    return env.LOCALE;
+}
+
+export async function loadLatestDataDragonVersion() {
+    if (env.DATADRAGON_VERSION !== 'latest') {
+        return;
+    }
+    const response = await fetch('https://ddragon.leagueoflegends.com/api/versions.json');
+    const versions = await response.json() as string[];
+    latestDataDragonVersion = versions[0];
+    console.log('Data Dragon version:', latestDataDragonVersion);
 }