+++ src/components/Match/MatchSummaryGroup.tsx
@@ -38,14 +38,10 @@ export default function MatchSummaryGroupComponent({ group, userIDs, report }: P
                                     <tr>
                                         <th>KDA</th>
                                         <td>
-                                            {group.KDA[id] && (
-                                                <FormattedNumber value={group.KDA[id]} />
-                                            )}
+                                            <FormattedNumber value={group.KDA[id]} />
                                         </td>
                                         <td>
-                                            {group.TotalKDA[id] && (
-                                                <FormattedNumber value={group.TotalKDA[id]} />
-                                            )}
+                                            <FormattedNumber value={group.TotalKDA[id]} />
                                         </td>
                                         <td>
                                             {group.PreviousGroup?.TotalKDA[id] && (
@@ -63,20 +59,16 @@ export default function MatchSummaryGroupComponent({ group, userIDs, report }: P
                                     <tr>
                                         <th>Win rate</th>
                                         <td>
-                                            {group.WinRate[id] && (
-                                                <FormattedNumber
-                                                    variant="percent"
-                                                    value={group.WinRate[id]}
-                                                />
-                                            )}
+                                            <FormattedNumber
+                                                variant="percent"
+                                                value={group.WinRate[id]}
+                                            />
                                         </td>
                                         <td>
-                                            {group.TotalWinRate[id] && (
-                                                <FormattedNumber
-                                                    variant="percent"
-                                                    value={group.TotalWinRate[id]}
-                                                />
-                                            )}
+                                            <FormattedNumber
+                                                variant="percent"
+                                                value={group.TotalWinRate[id]}
+                                            />
                                         </td>
                                         <td>
                                             {group.PreviousGroup?.TotalWinRate[id] && (
@@ -99,9 +91,11 @@ export default function MatchSummaryGroupComponent({ group, userIDs, report }: P
                 ))}
                 <td colSpan={1 + userIDs.length} />
             </tr>
-            {group.Matches.slice().reverse().map((match, i) => (
-                <MatchSummary key={match.ID} match={match} userIDs={userIDs} report={report} />
-            ))}
+            {group.Matches.slice()
+                .reverse()
+                .map((match, i) => (
+                    <MatchSummary key={match.ID} match={match} userIDs={userIDs} report={report} />
+                ))}
         </Fragment>
     );
 }