+++ src/components/Reports/MainReportContainer.tsx
@@ -1,0 +1,14 @@
+import { useMainReport } from '../../hooks/report';
+import StatsTable from '../Stats/StatsTable';
+
+export default function MainReportContainer() {
+    const report = useMainReport();
+    if (!report) {
+        return null;
+    }
+    return (
+        <div>
+            <StatsTable users={report.Users} />
+        </div>
+    );
+}