+++ src/components/Reports/MatchListReportContainer.tsx
@@ -1,0 +1,14 @@
+import { useMainReport } from '../../hooks/report';
+import MatchList from '../Match/MatchList';
+
+export default function MatchListReportContainer() {
+    const report = useMainReport();
+    if (!report) {
+        return null;
+    }
+    return (
+        <div>
+            <MatchList report={report} />
+        </div>
+    );
+}