+++ src/components/User/UserAvatar.tsx
@@ -3,12 +3,14 @@ import clsx from 'clsx';
 
 import { UserReport } from '../../types/common';
 import styles from './UserAvatar.module.css';
+import { useDataDragonVersion } from '../../hooks/report';
 
 interface Props {
     variant: 'heading';
     user: UserReport;
 }
 export default function UserAvatar({ variant, user }: Props) {
+    const dataDragonVersion = useDataDragonVersion();
     const [summonerIndex, setSummonerIndex] = useState(0);
     useEffect(() => {
         const interval = setInterval(() => {
@@ -27,7 +29,7 @@ export default function UserAvatar({ variant, user }: Props) {
                         [styles.selected]: summonerIndex === i,
                     })}
                 >
-                    <img alt="Icon" src={`/assets/${summoner.Icon}`} />
+                    <img alt="Icon" width={300} height={300} src={`/dd/${dataDragonVersion}/img/profileicon/${summoner.Icon}`} />
                     <code className={styles.level}>{summoner.Level}</code>
                     <div className={styles.name}>{summoner.Name}</div>
                 </div>