+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Match list
  - Riot API rate limiting and error handling
  - New awards (Trailblazer, Combat Medic, Dominator, Protector, Finisher, Pain Bringer, Siege Master, Creep Lover, Crowd Controller, Rich Bitch, Al-Qaeda, Pressing Intensifies)
+ - DATADRAGON_VERSION=latest & LOCALE env vars
 
 ## [0.1.0] - 2022-01-15
 ### Initial release